requirements.txt
streamlit    
python-dotenv
google-generativeai
requests
python-dotenv
pillow
psycopg2-binary
pypdf
google-generativeai
langchain
langchain-community
langchain-google-genai
chromadb